Modify the margin between chart lines
Icon images and data labels may overlap with each other
when the chart height is small (e.g. < 300px).

This commit enables to set the mergin based on
the chart height, font size and data labels offset values.

To calculate the margins, the "height" config must be in px.
Then I updated README to note that it must be in px.

Related issue:
https://github.com/mtatsuma/MMM-WeatherChart/issues/38